Move hera.h to public include dir
