+++ src/CMakeLists.txt
@@ -1,10 +1,13 @@
-get_filename_component(evm_include_dir .. ABSOLUTE)
-
 find_package(Threads REQUIRED)
 
+set(hera_include_dir ${PROJECT_SOURCE_DIR}/include)
+get_filename_component(evmc_include_dir .. ABSOLUTE)
+
 add_library(hera
-    eei.cpp eei.h
-    hera.cpp hera.h
+    ${hera_include_dir}/hera/hera.h
+    eei.cpp
+    eei.h
+    hera.cpp
 )
 
 option(HERA_DEBUGGING "Display debugging messages during execution." ON)
@@ -12,7 +15,7 @@ if(HERA_DEBUGGING)
   target_compile_definitions(hera PRIVATE HERA_DEBUGGING=1)
 endif()
 
-target_include_directories(hera PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} PRIVATE ${evm_include_dir})
+target_include_directories(hera PUBLIC ${hera_include_dir} PRIVATE ${evmc_include_dir})
 target_link_libraries(hera PRIVATE hera-buildinfo evmc::evmc evmc::instructions libevm2wasm binaryen::binaryen Threads::Threads)
 if(NOT WIN32)
   if(CMAKE_COMPILER_IS_GNUCXX)
